Moved the context menu implementation back to PropertiesWidget
This makes the property-specific actions like "Go to Object" and "Open
Containing Folder" available again, as well as the Cut, Copy and Paste
actions.

For now, the Custom Types Editor has no context menu anymore...

Property widgets are now removed before their property is deleted to
avoid a crash when right-clicking the AddValueProperty, since that
caused the focus to be lost and thus the removal of the property while
at the same time triggering a context menu for it.