Fixed tab order after creating property widgets
The tab order follows the widget creation order, which is not what we
want when properties are inserted between other properties. Hence we now
explicitly set the tab order in this case.

We can't just call QWidget::setTabOrder on the row widgets
unfortunately, because the call does nothing when called with widgets
that have Qt::NoFocus policy.

Calling QWidget::setTabOrder only on the inserted sequence led to other
issues, which led me to the solution of setting the tab order for all
property widgets to match the order in the layout. To avoid too much
repeated work when multiple properties are inserted, the fixTabOrder
call is scheduled using QMetaObject::invokeMethod.