AutoMapping: Removed the "touched layers" optimization
The "AutoMap While Drawing" had an optimization where it skipped entire
rule maps when none of their input layers had been "touched". This was
based on the assumption that its rules would not need to be re-evaluated
anyway.

Unfortunately, we can't make this assumption in general. The effect of
certain rules might be relevant even if none of their input or output
layers have been touched.