Fixed compile against Qt 6.8.2
In Qt 6.8.2, a forward declaration was added for the QStringRef class
from Qt 5, which breaks compilation because it is incompatible with the
"using QStringRef = QStringView" line I was using for compatibility
between Qt 5 and 6.

Fortunately, as of Qt 5.15.2, the QStringView API is complete enough, so
by raising the minimum required Qt version we no longer need QStringRef
at all.