libtiled-java: Fixed intermittent test failures
For some reason the resources failed to load from the `resources.jar`
file during testing when using `mvn clean test`. This also stood in the
way of deploying.

It seems to work reliably when generating the `resources.jar` directly
to the "target/test-classes" directory instead of using
"target/generated-test-resources" and then adding it with
build-helper-maven-plugin:add-test-resource.