libtiled-java: Fixed javadoc generation
The pegdown-doclet no longer works with recent Java versions.

Fortunately, we may not need it anymore anyway since it was "A Doclet
that allows the use of Markdown and PlantUML in JavaDoc comments" and
according to JavaDoc guide: "The standard doclet for the javadoc tool
supports the use of the CommonMark variant of Markdown in documentation
comments".