Fixed crash on macOS when JS code is JIT-compiled
The codesigned release did not have the right entitlement to be able to
execute JIT-compiled code.

Also, the MAP_JIT memory flag is only set by Qt since 6.8.2, so when
Tiled is compiled against earlier Qt versions, we now force the use of
the JS interpreter to disable the JIT compiler.

TODO: This change relies on the Qbs codesign module (available since Qbs
1.19), which will likely need CI updates.

Closes #4218