Allow JS expressions in number inputs (#4234)
Replaced QSpinBox and QDoubleSpinBox throughout the app, including in
script dialogs and properties view, with subclasses that allows setting
a value using a JavaScript expression.

Any prefix or suffix is temporarily hidden while a spin box has focus 
because it gets in the way or is automatically added when it shouldn't.
This way we also don't need to parse them out before evaluating the
text.

QSpinBox::setValue and QDoubleSpinBox::setValue cause the text to get
reset to the current value, but we don't want to do this while the user
is still typing. So in the Properties view we now compare values first
to not call setValue needlessly.

We don't evaluate the expression in the validate implementation because
this is called a bit too much to be executing JS. It also gets called
when plain number values are set on the spin box, for example during
initialization.

Co-authored-by: Thorbjørn Lindeijer <bjorn@lindeijer.nl>