Added additional description to Stagger Axis property
A staggered X axis leads to Flat-Topped hexes whereas a staggered Y axis
leads to Pointy-Topped hexes.

This setting also applies to maps using Isometric (Staggered) orientation,
in which case this distinction does not exist, so I'm not entirely sure
whether this is a good idea. Should we have the property only use the
additional labels when the map is using Hexagonal orientation?