Fixed possible crash in the Custom Types Editor
Due to a missing parent on the GroupProperty used by the
PropertiesWidget, the GroupProperty instance would not get deleted,
which in turns causes the CustomProperties instance to not get deleted.

The leaked CustomProperties instance would later cause a crash when
editing the custom types because it accesses a roaming Document pointer
while it tries to update itself.