Prevent creation of cyclic references in custom types
When adding values to lists that are part of a custom class definition,
it was possible to create a cyclic reference. Now the same mechanism
that was used to prevent such references when adding members is also
applied when adding list values.

Since lists can be nested this was a little more involved. The
PropertyTypesEditor now uses a custom PropertiesView subclass which the
VariantListProperty and AddValueProperty now use to obtain the currently
edited class type. To enable this, each Property which is shown in a
PropertiesView now gets a reference to this view.