Fix issue with opening Tile Animation Editor (#4264)
On Windows, with Tiled compiled against Qt 6, the Tile Animation Editor
would fail to become visible when it had been left in a maximized state
before closing Tiled.

Somehow the dialog in a maximized state was receiving WM_DESTROY when
its parent QMainWindow was added to the layout in
DocumentManager::setEditor:

> External WM_DESTROY received for QWidgetWindow(0x20d997eba40, name="TileAnimationEditorWindow") , parent: QWindow(0x0) , transient parent: QWindow(0x0)

I'm not entirely sure what causes this behavior, but I've found that
lazily creating the TileAnimationEditor on the first time it is opened
provided a workaround.

Closes #4223