Add named properties on focus lost (#4265)
When adding a new property or class member, losing the focus of the "add
value widget" always aborted the addition. Now, it instead adds the
property when a name had been entered.

In this case, the value edit for the new property is not explicitly
focused, to avoid disrupting the user's focus change.

A few direct signal-to-signal connections were exchanged for lambdas to
avoid the "focus" parameter getting set based on a "checked" parameter
in the original signals.