Adjust world map position when resizing a map with offset
When an offset is applied while resizing a map, this offset will now
affect the maps position in any loaded worlds the map is in. This
affects both manual resizing as well as the "crop to selection" and
"autocrop" actions.

Implemented using a SetMapPosInLoadedWorld command which has a weak
reference (file name) to a loaded world, to avoid issues when a world is
later unloaded. It in turn uses SetMapRectCommand to apply the change to
the world, which was adjusted to set itself as obsolete when applicable
in order to disappear when the resize operation is undone.

Closes #4270