Added support for natural sorting of project files (#4284)
This change adds a natural / numeric sorting option affecting Projects 
view and the results in Open File in Project, which is enabled by
default.

It is implemented using a proxy model to enable adjusting the Project 
view quickly when the option is changed, without needing to rescan the
folders.

Also ensured `.vscode` config dirs are git ignored.

Closes #4180

Co-authored-by: Thorbjørn Lindeijer <bjorn@lindeijer.nl>